refactor(web): restructure header component location
Moved header component to a dedicated folder for better organization.
 Adjusted import paths accordingly to reflect the new component
 hierarchy.
